requirements.txt
dune-client
python-dotenv 
pandas
inquirer